Splitting the rotation Algos out of the hexAssemUtils
